add test for er as stim and fix non-ascii bug
